apps/agent-host/pyproject.toml
[tool.poetry]
name = "sql-agent-oss"
version = "0.1.0"
description = "Agente Text-to-SQL Open Source con arquitectura semántica y autofix"
authors = ["David R. <dev.drbh@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"

# --- Núcleo IA y Agentes ---
langchain = "^1.2.0"
langgraph = "^1.0.5"

# --- Base de Datos (Async) ---
sqlalchemy = "^2.0.25"
# Driver para MySQL Async (Vital para tu copia de producción)
aiomysql = "^0.2.0"
cryptography = "^42.0.0" # Necesario para autenticación en MySQL 8

# Driver Postgres (Lo dejamos por si acaso, no estorba)
asyncpg = "^0.29.0"
psycopg = {extras = ["binary", "pool"], version = "^3.2.4"}
langgraph-checkpoint-postgres = "^2.0.9"

# --- Manejo de Datos y Fuzzy Search ---
pandas = "^2.2.0"           
thefuzz = "^0.22.0"         
python-Levenshtein = "^0.25.0" 

# --- Base de Datos Vectorial ---
chromadb = "^0.4.22"

# --- Seguridad ---
sqlglot = "^20.8.0"

# --- UI ---
chainlit = "^2.9.4"

# --- Utilidades ---
python-dotenv = "^1.0.0"
pydantic = "^2.6.0"
pyyaml = "^6.0.1"
langchain-google-genai = "^4.1.3"
google-generativeai = "^0.8.6"
langchain-openai = "^1.1.6"
langchain-community = "^0.4.1"
tiktoken = "^0.12.0"
fastapi = "^0.128.0"
uvicorn = "^0.40.0"
requests = "^2.32.5"
aiohttp = "^3.13.3"
mcp = "^1.25.0"

# ACTUALIZADO: httpx debe ser >=0.28.1 para compatibilidad con google-genai
httpx = "^0.28.1"
langchain-mcp-adapters = "^0.2.1"
python-telegram-bot = "^22.5"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.14"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
testcontainers = "^3.7.1"
mkdocs = "1.6.0"
mkdocs-material = "^9.7.1"
griffe = "^1.15.0"
mkdocstrings = {extras = ["python"], version = "^1.0.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"